mac/dialog: move dialog creation into its own class
